+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.7.4 - 2014-11-12
 
+* Added support for the AWS Key Management Service.
 * Added support for AWS CodeDeploy.
 * Added support for Amazon KMS encryption to the Amazon EC2 client.
 * Added support for Amazon CloudWatch Logs delivery to the AWS CloudTrail